Split `iter::intervals::layered_bounds`
