+++ src/iter/intervals.rs
@@ -28,7 +28,6 @@ pub mod united_bounds;
 tests! {
     mod bounds_tests;
     mod complement_tests;
-    mod layered_bounds_tests;
     mod remove_empty_tests;
     // mod split_tests;
     mod united_bounds_tests;
