+++ src/iter/intervals/layered_bounds/abs_state_change_tests.rs
@@ -1,0 +1,86 @@
+use std::error::Error;
+
+use jiff::Zoned;
+
+use super::abs_state_change::*;
+use crate::intervals::absolute::{AbsoluteEndBound, AbsoluteFiniteBound, AbsoluteStartBound};
+use crate::intervals::meta::BoundInclusivity;
+use crate::iter::intervals::layered_bounds::state::LayeredBoundsState;
+
+#[test]
+fn at_abs_bound_old_state() {
+    assert_eq!(
+        LayeredBoundsStateChangeAtAbsoluteBound::new(
+            LayeredBoundsState::FirstLayer,
+            LayeredBoundsState::SecondLayer,
+            None,
+            None,
+        )
+        .old_state(),
+        LayeredBoundsState::FirstLayer,
+    );
+}
+
+#[test]
+fn at_abs_bound_new_state() {
+    assert_eq!(
+        LayeredBoundsStateChangeAtAbsoluteBound::new(
+            LayeredBoundsState::FirstLayer,
+            LayeredBoundsState::SecondLayer,
+            None,
+            None,
+        )
+        .new_state(),
+        LayeredBoundsState::SecondLayer,
+    );
+}
+
+#[test]
+fn at_abs_bound_old_state_end() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        LayeredBoundsStateChangeAtAbsoluteBound::new(
+            LayeredBoundsState::FirstLayer,
+            LayeredBoundsState::SecondLayer,
+            Some(AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            ))),
+            Some(AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            ))),
+        )
+        .old_state_end(),
+        Some(AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+            "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+            BoundInclusivity::Inclusive,
+        ))),
+    );
+
+    Ok(())
+}
+
+#[test]
+fn at_abs_bound_new_state_start() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        LayeredBoundsStateChangeAtAbsoluteBound::new(
+            LayeredBoundsState::FirstLayer,
+            LayeredBoundsState::SecondLayer,
+            Some(AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            ))),
+            Some(AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            ))),
+        )
+        .new_state_start(),
+        Some(AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+            "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+            BoundInclusivity::Exclusive,
+        ))),
+    );
+
+    Ok(())
+}
